Silo layer-specific types into the layers directory.
