+++ src/core/index.ts
@@ -3,5 +3,5 @@ export * from './intervalUtils';
 export * from './componentUtils';
 export * from './decorators';
 export * from './interfaces';
-export { default as Stack } from './Stack';
 export { default as propTypes } from './propTypes';
+export { default as Stack } from './Stack';
