+++ src/layers/connected/ConnectedBrushLayer.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import * as PureRender from 'pure-render-decorator';
 import { connect } from 'react-redux';
 
-import { Interval, Color, XSpanDatum } from '../../core';
+import { Interval, Color } from '../../core';
 import { ChartProviderState, selectSelection, selectXDomain } from '../../connected';
 import BrushLayer from '../BrushLayer';
