+++ src/layers/index.ts
@@ -1,3 +1,6 @@
+export * from './layerDataTypes';
+export * from './connected';
+
 export {
   default as AutoresizingCanvasLayer,
   Props as AutoresizingCanvasLayerProps
@@ -57,5 +60,3 @@ export {
   default as YAxisLayer,
   Props as YAxisLayerProps
 } from './YAxisLayer';
-
-export * from './connected';
