+++ test/layers/BarLayer-test.ts
@@ -4,7 +4,7 @@ import * as d3Scale from 'd3-scale';
 
 import { span, method } from './layerTestUtils';
 import CanvasContextSpy from '../../src/test-util/CanvasContextSpy';
-import { SpanDatum } from '../../src/core/interfaces';
+import { SpanDatum } from '../../src/layers/layerDataTypes';
 import { _renderCanvas, Props } from '../../src/layers/BarLayer';
 
 describe('BarLayer', () => {
