+++ test/layers/SpanLayer-test.ts
@@ -4,7 +4,7 @@ import * as d3Scale from 'd3-scale';
 
 import { method, property, xSpan } from './layerTestUtils';
 import CanvasContextSpy from '../../src/test-util/CanvasContextSpy';
-import { XSpanDatum } from '../../src/core/interfaces';
+import { XSpanDatum } from '../../src/layers/layerDataTypes';
 import { _renderCanvas, Props } from '../../src/layers/SpanLayer';
 
 describe('SpanLayer', () => {
